+++ app/build.gradle
@@ -27,4 +27,6 @@ dependencies {
     compile 'com.mobilesolutionworks:works-util:1.0.4'
     compile 'com.android.support:cardview-v7:23.0.1'
     compile 'com.android.support:recyclerview-v7:23.0.1'
+    compile 'fr.nelaupe:navigation-fragment-handler:1.0.3@aar'
+
 }
