+++ app/src/main/java/com/nelaupe/qanda/entity/Answer.java
@@ -14,5 +14,4 @@ public class Answer {
 
     public String content;
 
-
 }
