+++ app/src/main/java/com/nelaupe/qanda/fragment/AnswerFragment.java
@@ -1,9 +1,8 @@
 /**
  * Copyright
  */
-package com.nelaupe.qanda;
+package com.nelaupe.qanda.fragment;
 
-import android.app.Fragment;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.widget.LinearLayoutManager;
@@ -13,6 +12,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import com.nelaupe.qanda.R;
 import com.nelaupe.qanda.entity.Answer;
 
 import java.util.ArrayList;
@@ -23,7 +23,7 @@ import java.util.List;
  * Created by lucas
  * Date 26/03/15
  */
-public class AnswerFragment extends Fragment {
+public class AnswerFragment extends BaseFragment {
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -33,7 +33,7 @@ public class AnswerFragment extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.answer_layout, container, false);
+        return inflater.inflate(R.layout.answer_fragment, container, false);
     }
 
     @Override
