+++ app/src/main/res/layout/answer_fragment.xml
